requirements.txt
Django==1.10.1
django-formtools==1.0
Pillow==3.3.1
psycopg2==2.6.2
